change build to use 'dep' (again)
Struggle with the stupidity of building and testing Go programs with this.  had some issue with Glide not being able to download the 'go-check' package, so I switched to 'dep' .

In addition, i also configured the wercker build to use some stupid bash-ism to list the files to build, as Go's build tool is not smart enough to ignore or skip testing the dependencies.